CanvasHs.Data

Documentation

type Point = (Int, Int)

type Path = [Point]

type Color = (Int, Int, Int, Int)

class Defaults a where

Methods

defaults :: a

data EventData

Constructors

EventData

Fields

eventId :: String

The ID of the event, should be unique.

mouseDown :: Bool

Toggles whether to react on mouseDown.

mouseClick :: Bool

Toggles whether to react on mouseClick.

mouseUp :: Bool

Toggles whether to react on mouseUp.

mouseDoubleClick :: Bool

Toggles whether to react on mouseDoubleClick.

mouseDrag :: Bool

Toggles whether to react on mouseDrag.

mouseEnter :: Bool

Toggles whether to react on mouseEnter.

mouseLeave :: Bool

Toggles whether to react on mouseLeave.

data FontSize

Constructors

Pt Int

Px Int

data Alignment

Constructors

Start
Aligns the start of the text to the specified point.

End
Aligns the end of the text to the specified point.

Center
Aligns the center of the text to the speciied point.

data TextData

Constructors

TextData

Fields

font :: String

The font for this text, no guarantees are made about availability

size :: FontSize

The fontsize in Points or Pixels

bold :: Bool

Toggles bold.

italic :: Bool

Toggles italic.

underline :: Bool

Toggles underline.

alignment :: Alignment

Specifies how to align this text.

data Shape

Constructors

Rect Point Int Int
A rectangle. Has a startpoint (left upper corner) and width, height

Circle Point Int
A circle. Has a centerpoint and a radius.

Arc Point Int Int Int
An arch. Has a centerpoint, radius, startangle and endangle (counterclockwise).

Line Path
A line. Has a path containing its points, doesn't connect start and end.

Polygon Path
A polygon. Has a path containing its points, does connect start with end.

Text Point String TextData
A text. Has a centerpoint a string containing the text and some extra data.

Fill Color Shape
Applies fill. Has a color and a shape that needs to be filled.s

Stroke Color Int Shape
Applies stroke. Has a color, a strokewidth and a shape that needs to be stroked.

Rotate Int Shape
Applies rotate. Has a rotation (counterclockwise) and a shape that needs to be rotated.

Translate Int Int Shape
Applies translation. Has xdiff, ydiff and a shape that needs to be translated.

Scale Float Float Shape
Applies scale. Has xscale, yscale and a shape that needs to be scaled.

Event EventData Shape

Container Int Int [Shape]
A container. Has width and height and a list of shapes in this container.