CanvasHs.Data

Documentation

type Point = (Int, Int)

Convenience type for a point (x, y)

type Path = [Point]

Convenience type for a path consisting of points

type Color = (Int, Int, Int, Int)

Convenience type for a RGBA color

class Defaults a where

A class that defines a function defaults that returns a default value

Methods

defaults :: a

data EventData

Stores data about what events a shape is interested in

Constructors

EventData

Fields

eventId :: String

The ID of the event, should be unique.

mouseDown :: Bool

Toggles whether to react on mouseDown.

mouseClick :: Bool

Toggles whether to react on mouseClick.

mouseUp :: Bool

Toggles whether to react on mouseUp.

mouseDoubleClick :: Bool

Toggles whether to react on mouseDoubleClick.

mouseDrag :: Bool

Toggles whether to react on mouseDrag.

mouseOver :: Bool

Toggles whether to react on mouseOver

mouseOut :: Bool

Toggles whether to react on mouseOut.

data FontSize

Defines textsize

Constructors

Pt Int

Px Int

data Alignment

Ways to align text

Constructors

Start
Aligns the start of the text to the specified point.

End
Aligns the end of the text to the specified point.

Center
Aligns the center of the text to the speciied point.

data TextData

Record holding formatting data for text, eg fonts and sizes

Constructors

TextData

Fields

font :: String

The font for this text, no guarantees are made about availability

size :: FontSize

The fontsize in Points or Pixels

bold :: Bool

Toggles bold.

italic :: Bool

Toggles italic.

underline :: Bool

Toggles underline.

alignment :: Alignment

Specifies how to align this text.

data Shape

All drawable objects that the user can define (also includes some modifying objects like scale)

Constructors

Rect Point Int Int
A rectangle. Has a startpoint (left upper corner) and width, height

Circle Point Int
A circle. Has a centerpoint and a radius.

Arc Point Int Int Int
An arch. Has a centerpoint, radius, startangle and endangle (counterclockwise).\ | Not yet implemented

Line Path
A line. Has a path containing its points, doesn't connect start and end.

Polygon Path
A polygon. Has a path containing its points, does connect start with end.

Text Point String TextData
A text. Has a centerpoint a string containing the text and some extra data.

Fill Color Shape
Applies fill. Has a color and a shape that needs to be filled.

Stroke Color Int Shape
Applies stroke. Has a color, a strokewidth and a shape that needs to be stroked.

Rotate Int Shape
Applies rotate. Has a rotation (counterclockwise) and a shape that needs to be rotated.

Translate Int Int Shape
Applies translation. Has xdiff, ydiff and a shape that needs to be translated.

Scale Float Float Shape
Applies scale. Has xscale, yscale and a shape that needs to be scaled.

Event EventData Shape

Container Int Int [Shape]
A container. Has width and height and a list of shapes in this container.

data Modifier

Keymodifiers that can be enabled in a keyboard event

Constructors

Shift

Ctrl

Alt

Super

data Event

The events the user can expect to get as input

Constructors

MouseDown Point String
A mousedown event consisting of a point and ID string of the interested object

MouseClick Point String
A mouseclick event consisting of a point and ID string of the interested object

MouseUp Point String
A mouseup event consisting of a point and ID string of the interested object

MouseDoubleClick Point String
A mousedoubleclick event consisting of a point and ID string of the interested object

MouseDrag Point String Point String
A mousedrag event with start and end, both consisting of a Point an ID string

MouseOver Point String
A mouseover event, therefore both Point and an ID string are set.

MouseOut Point String
A mouseout event, therefore both Point and an ID string are set.

KeyDown Char [Modifier]
A keydown event, consist of a keycharacter that was pressed and a list of modifiers that were active

KeyUp Char [Modifier]
A keyup event, consist of a keycharacter that was pressed and a list of modifiers that were active

KeyClick Char [Modifier]
A keyclick event, consist of a keycharacter that was pressed and a list of modifiers that were active

Scroll Int Int
A scroll event consisting of a xdiff and a ydiff

StartEvent
Start event is thrown when the server is started to notify user